requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
apache-libcloud==3.8.0
async==0.6.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.13.1
bcrypt==4.1.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fasteners==0.19
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.4
redis-dict==2.5.0
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlalchemy-file==0.6.0
starlette==0.37.2
starlette-admin==0.13.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
